+++ Braintree/CustomerGateway.cs
@@ -55,18 +55,16 @@ namespace Braintree
             return Configuration.BaseMerchantURL() + "/customers/all/update_via_transparent_redirect_request";
         }
 
-        public PagedCollection<Customer> All()
+        public ResourceCollection<Customer> All()
         {
             return All(1);
         }
 
-        public PagedCollection<Customer> All(Int32 pageNumber)
+        public ResourceCollection<Customer> All(Int32 pageNumber)
         {
             String queryString = new QueryString().Append("page", pageNumber).ToString();
             NodeWrapper response = new NodeWrapper(WebServiceGateway.Get("/customers?" + queryString));
             
-            int currentPageNumber = response.GetInteger("current-page-number").Value;
-            int pageSize = response.GetInteger("page-size").Value;
             int totalItems = response.GetInteger("total-items").Value;
 
             List<Customer> customers = new List<Customer>();
@@ -75,7 +73,7 @@ namespace Braintree
                 customers.Add(new Customer(node));
             }
 
-            return new PagedCollection<Customer>(customers, currentPageNumber, totalItems, pageSize, delegate() { return All(pageNumber + 1); });
+            return new ResourceCollection<Customer>(customers, totalItems, delegate() { return All(pageNumber + 1); });
         }
     }
 }